requirements.txt
alembic==1.13.1
black==24.2.0
click==8.1.7
colorama==0.4.6
flake8==7.0.0
greenlet==3.0.3
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pep8==1.7.1
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
python-decouple==3.8
rich==13.7.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
